+++ apps/client/src/components/features/main/Banner.tsx
